refactor(tokenstore): add ts
